move EventHandlerUtils
